p14: moving from u64 to u32 shaves off about 15%, so around 7ms.

+++ src/main.rs
@@ -429,7 +429,7 @@ fn p13() -> u64 {
 }
 
 #[allow(dead_code)]
-fn p14() -> usize {
+fn p14() -> u32 {
     eulerrust::collatz::longest_collatz_memo(999_999)
 }
